Added new function in function.js, created new modules, adjustments have been made
